+++ tests/browser_interactions.test.js
@@ -1,5 +1,6 @@
 "use strict";
 
+const assert = require('assert');
 const Wendigo = require('../lib/wendigo');
 const utils = require('./utils');
 const configUrls = require('./config.json').urls;
@@ -19,14 +20,16 @@ describe("Browser Interactions", function() {
     it("Click", async() => {
         await browser.open(configUrls.click);
         await browser.assert.text("#switch", "On");
-        await browser.click(".btn");
+        const clickedElements = await browser.click(".btn");
         await browser.assert.text("#switch", "Off");
+        assert.strictEqual(clickedElements, 1);
     });
 
     it("Click Multiple Elements", async() => {
         await browser.open(configUrls.click);
         await browser.assert.text("#switch", "On");
-        await browser.click("button");
+        const clickedElements = await browser.click("button");
+        assert.strictEqual(clickedElements, 2);
         await browser.assert.text("#switch", "Off");
         await browser.waitFor("#switch.on", 600);
     });
@@ -34,7 +37,8 @@ describe("Browser Interactions", function() {
     it("Click With Index", async() => {
         await browser.open(configUrls.click);
         await browser.assert.text("#switch", "On");
-        await browser.click("button", 1);
+        const clickedElements = await browser.click("button", 1);
+        assert.strictEqual(clickedElements, 1);
         await browser.assert.text("#switch", "On");
         await browser.waitFor("#switch.off", 600);
     });
@@ -42,7 +46,8 @@ describe("Browser Interactions", function() {
     it("Click From Node", async() => {
         await browser.open(configUrls.click);
         const node = await browser.query(".btn");
-        await browser.click(node);
+        const clickedElements = await browser.click(node);
+        assert.strictEqual(clickedElements, 1);
         await browser.assert.text("#switch", "Off");
     });
 
@@ -51,7 +56,7 @@ describe("Browser Interactions", function() {
         await browser.assert.text("#switch", "On");
         await utils.assertThrowsAsync(async () => {
             await browser.click("button", 10);
-        }); // TODO: check text
+        }, `Error: browser.click, invalid index "10" for selector "button", 2 elements found.`);
         await browser.assert.text("#switch", "On");
     });
 
@@ -88,7 +93,7 @@ describe("Browser Interactions", function() {
         await browser.click(".btn2");
         await utils.assertThrowsAsync(async () => {
             await browser.waitFor("#switch.off", 10);
-        });
+        }, `Error: Waiting for element "#switch.off" failed, timeout of 10ms exceeded`);
         await browser.assert.not.exists("#switch.off");
         await browser.assert.exists("#switch.on");
         await browser.assert.text("#switch", "On");
@@ -96,7 +101,8 @@ describe("Browser Interactions", function() {
 
     it("Click Text", async() => {
         await browser.open(configUrls.click);
-        await browser.clickText("click me");
+        const clickedElements = await browser.clickText("click me");
+        assert.strictEqual(clickedElements, 1);
         await browser.assert.text("#switch", "Off");
         await browser.clickText("click me");
         await browser.assert.text("#switch", "On");
@@ -104,7 +110,9 @@ describe("Browser Interactions", function() {
 
     it("Click Invalid Text", async() => {
         await browser.open(configUrls.click);
-        await browser.clickText("not click me");
+        await utils.assertThrowsAsync(async () => {
+            await browser.clickText("not click me");
+        }, `Error: No element with text "not click me" found when trying to click.`);
         await browser.assert.text("#switch", "On");
     });
 
@@ -115,4 +123,6 @@ describe("Browser Interactions", function() {
         await browser.assert.text("#switch", "Off");
         await browser.waitFor("#switch.on", 600);
     });
+
+    it("Click Invalid Element");
 });